backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.5
sentence-transformers==2.2.2
chromadb==0.4.18
python-dotenv==1.0.0
spacy==3.7.2
dateutil==2.8.2
transformers==4.35.2
torch==2.1.1